requirements.txt
-r requirements/optional.txt
-r requirements/runtime.txt
-r requirements/tests.txt
cuda                       11.3.0
addict                     2.4.0            pypi_0           pypi
ca-certificates            2025.9.9         haa95532_0       https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
clip-anytorch              2.6.0            pypi_0           pypi
colorama                   0.4.6            pypi_0           pypi
contourpy                  1.1.1            pypi_0           pypi
cycler                     0.12.1           pypi_0           pypi
fonttools                  4.57.0           pypi_0           pypi
ftfy                       6.2.0            pypi_0           pypi
imageio                    2.35.1           pypi_0           pypi
importlib-resources        6.4.5            pypi_0           pypi
kiwisolver                 1.4.7            pypi_0           pypi
lazy-loader                0.4              pypi_0           pypi
matplotlib                 3.7.5            pypi_0           pypi
mmcv-full                  1.4.0            pypi_0           pypi
mmsegmentation             0.20.2           pypi_0           pypi
networkx                   3.1              pypi_0           pypi
numpy                      1.24.4           pypi_0           pypi
opencv-python              4.12.0.88        pypi_0           pypi
openssl                    1.1.1w           h2bbff1b_0       https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
packaging                  25.0             pypi_0           pypi
pillow                     10.4.0           pypi_0           pypi
pip                        24.2             py38haa95532_0   https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
platformdirs               4.3.6            pypi_0           pypi
prettytable                3.11.0           pypi_0           pypi
pydensecrf                 1.0rc2           pypi_0           pypi
pyparsing                  3.1.4            pypi_0           pypi
python                     3.8.0            hff0d562_2       https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
python-dateutil            2.9.0.post0      pypi_0           pypi
pywavelets                 1.4.1            pypi_0           pypi
pyyaml                     6.0.3            pypi_0           pypi
regex                      2024.11.6        pypi_0           pypi
scikit-image               0.21.0           pypi_0           pypi
scipy                      1.10.1           pypi_0           pypi
setuptools                 75.1.0           py38haa95532_0   https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
six                        1.17.0           pypi_0           pypi
sqlite                     3.50.2           hda9a48d_1       https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
tifffile                   2023.7.10        pypi_0           pypi
tomli                      2.3.0            pypi_0           pypi
torch                      1.10.0+cu113     pypi_0           pypi
torchaudio                 0.10.0+cu113     pypi_0           pypi
torchvision                0.11.0+cu113     pypi_0           pypi
tqdm                       4.67.1           pypi_0           pypi
typing-extensions          4.13.2           pypi_0           pypi
ucrt                       10.0.22621.0     haa95532_0       https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
vc                         14.3             h2df5915_10      https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
vc14_runtime               14.44.35208      h4927774_10      https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
vs2015_runtime             14.44.35208      ha6b5a95_10      https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
wand                       0.6.13           pypi_0           pypi
wcwidth                    0.2.14           pypi_0           pypi
wheel                      0.44.0           py38haa95532_0   https://mirrors.tuna.tsinghua.edu.cn/anaconda/pkgs/main
yapf                       0.30.0           pypi_0           pypi
zipp                       3.20.2           pypi_0           pypi